refactor: consolidate subcomponents into the same files
